Automate i64 -> i53 conversions parameters and return values of JS library funcs.
Previously folks would need to use helper macros like `defineI64Param`
and `receiveI64ParamAsI53`, but with this change that is now completely
automatic, based on the signature (`__sig`) of the function.

I've kept the old helper function around in `parseTools_legacy.js` in
case there are external folks using them.

There may be some cases where this automatic conversion is not desirable
and a function is written to accept actual BigInt values (that represent
the full i64 range).  In indeed it turns out there once example of such
as function within emscripten itself (`embind_register_bigint`).  In
order to mark such function I have created a new `__bigint` attribute
that can be used on JS functions.  I imagine that the need for this
attribute will be very rare.